fix(Notification): change chat id type

+++ app/Listeners/Notification/OrderNotification.php
@@ -18,7 +18,7 @@ class OrderNotification
 
         $this->notificationService->setService(new TelegramNotification(
             config('services.telegram.token'),
-            config('services.telegram.chat_id')
+            (int) config('services.telegram.chat_id')
         ));
 
         $this->notificationService->notify($message, 'Новый заказ');
